Loads of improvements, sorry should have updated earlier...
- Removed ROM loader browser, not useful anymore.
- Changed Menu to the one used in SMS Plus GX fork.
- Fixed audio backends for ALSA, portaudio, and OSS.
- Added back EEPROM and Save states to Menu.

Each game has its own configuration file.
By default, they use a bunch of preset input that is okay for most games
out there but you might want to configure them to your likings.
Said config file also saves the game's orientation.
When it can be detected (via header or using the appropriate IO port for that),
the game automatically switches to being rotated unless told otherwise.

In some cases (Klonoa), this cannot be detected in which case
you'll have to trigger it manually.

You may also want to play the game vertically, which is possible
if you have a monitor that has a rotable screen or a handled.
In this case, you'll have to disable rotation and change the controls
in vertical mode.

As for the video specific code files, vertical games are now rotated.
(Counter clock wise only. Not aware of any games supporting clock wise
only rotation)

Input code also had to be changed accordingly for rotation :
nothing too tricky or fancy, it was fairly simple.